AUTO_WEBSITE/requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
celery==5.4.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.7
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==5.0.1
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==23.5
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
hiredis==2.3.2
idna==3.6
kombu==5.3.7
multidict==6.0.5
mysql-connector-python==8.3.0
mysqlclient==2.2.1
oauthlib==3.2.2
phonenumbers==8.13.28
pillow==10.2.0
prompt-toolkit==3.0.43
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2023.3.post1
redis==5.0.4
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
twilio==8.13.0
typing_extensions==4.11.0
tzdata==2023.4
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
yarl==1.9.4